gamesrouter tests

+++ routes/sportRouter.js
@@ -1,4 +1,5 @@
 const express = require('express')
+const config = require('../utils/config')
 const Sport = require('../models/sport')
 const router = express.Router()
 
@@ -17,11 +18,12 @@ router.post('/', async (req, res) => {
   const name = data['name']
 
   const newSport = Sport({ name: name })
-  await newSport.save()
-  res.status(201).end()
+  const saved = await newSport.save()
+  res.status(201).json(saved).end()
 })
 
 router.delete('/', async (req, res) => {
+  if (config.NODE_ENV !== 'test') return
   await Sport.deleteMany({})
   res.status(204).end()
 })
